backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-utils-six==2.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
Pillow==9.3.0
psycopg2-binary==2.9.1
pytz==2022.6
PyYAML==6.0
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12

